Write out the script to initialize `home-manager`
Apparently, it doesn't work to clone the private repo in the
`install.sh` script, despite that being the recommendation we've heard
from the Workspaces folks. Given how unintuitive this is, it's very
likely that I misunderstood the recommendation.

We change tactics and move a script to `~/.local/bin` (which should be
on the path). This script does the initialization of `home-manager`.
It'd be nice to have this happen immediately, but what can you do.