Set the `--backup-extension`
If we don't do this, the `home-manager` activation script won't actually
run. Worse still, it will silently fail without saying _why_ it didn't
run. So we think it did succeed.

We don't care about any previous configuration, since these workspaces
are ephemeral, and the whole point is to use `home-manager` to manage
the configuration.